@@ -15,48 +15,56 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "^3.1.0",
-    "@docusaurus/plugin-content-docs": "^3.1.0",
-    "@docusaurus/plugin-ideal-image": "^3.1.0",
-    "@docusaurus/preset-classic": "^3.1.0",
-    "@docusaurus/remark-plugin-npm2yarn": "^3.1.0",
-    "@docusaurus/theme-common": "^3.1.0",
-    "@docusaurus/utils": "^3.1.0",
-    "@fortawesome/fontawesome-svg-core": "^6.5.1",
-    "@fortawesome/free-brands-svg-icons": "^6.5.1",
-    "@fortawesome/free-regular-svg-icons": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^6.5.1",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mdx-js/react": "^3.0.0",
+    "@docusaurus/core": "^3.5.0",
+    "@docusaurus/plugin-content-docs": "^3.5.0",
+    "@docusaurus/plugin-ideal-image": "^3.5.0",
+    "@docusaurus/preset-classic": "^3.5.0",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.5.0",
+    "@docusaurus/theme-classic": "^3.5.0",
+    "@docusaurus/theme-common": "^3.5.0",
+    "@docusaurus/utils": "^3.5.0",
+    "@easyops-cn/docusaurus-search-local": "^0.44.4",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-regular-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@iconify/react": "^5.0.2",
+    "@mdx-js/react": "^3.0.1",
     "@popperjs/core": "^2.11.8",
-    "@swc/core": "^1.3.102",
-    "@types/react-dom": "^18.2.18",
-    "clsx": "^2.1.0",
+    "@swc/core": "^1.7.9",
+    "@types/react-dom": "^18.3.0",
+    "clsx": "^2.1.1",
     "docusaurus-plugin-plausible": "^0.0.5",
     "docusaurus-plugin-sass": "^0.2.5",
-    "docusaurus-remark-plugin-tab-blocks": "^3.0.0",
+    "docusaurus-remark-plugin-tab-blocks": "^3.1.0",
+    "git-url-parse": "^14.1.0",
     "prism-react-renderer": "^2.3.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-ideal-image": "^0.0.5",
     "react-popper": "^2.3.0",
     "rehype-katex": "7",
     "remark-definition-list": "^2.0.0",
     "remark-deflist": "^1.0.0",
     "remark-math": "6",
-    "sass": "^1.69.7",
-    "swc-loader": "^0.2.3"
+    "sass": "^1.77.1",
+    "swc-loader": "^0.2.6"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^3.1.0",
-    "@docusaurus/tsconfig": "^3.1.0",
-    "@types/jest": "^29.5.11",
-    "@types/react": "^18.2.47",
+    "@docusaurus/module-type-aliases": "^3.5.0",
+    "@docusaurus/tsconfig": "^3.5.0",
+    "@docusaurus/types": "^3.5.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^22.1.0",
+    "@types/react": "^18.3.3",
     "jest": "^29.7.0",
+    "prettier": "3.3.3",
+    "prettier-plugin-css-order": "2.1.2",
     "prop-types": "^15.8.1",
     "react-waypoint": "^10.3.0",
-    "typescript": "^5.3.3",
-    "webpack": "^5.89.0"
+    "typescript": "^5.5.4",
+    "webpack": "^5.93.0"
   },
   "browserslist": {
     "production": [
@@ -73,5 +81,10 @@
   "engines": {
     "node": ">=18.14"
   },
-  "packageManager": "yarn@4.0.2"
+  "packageManager": "yarn@4.4.0",
+  "dependenciesMeta": {
+    "@docusaurus/tsconfig": {
+      "unplugged": true
+    }
+  }
 }
